[Get Most Populated Countries] Calculate Density and Sort

+++ restcountries-model/src/main/java/com/furqonajiy/restcountries/model/backend/restcountries/RestCountriesResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class RestCountriesResponse {
     private Name name;
-    private Integer area;
-    private Integer population;
+    private double area;
+    private int population;
 }
